# TRY perf(http): zero copy response body bytes