`Response` clones the buffer JS side. `Bytes::copy_from_slice` did a second copy of the same data.

```
deno run -A https://deno.land/std/http/file_server.ts
```

`main`:
```
$ wrk -t8 -c20 -d10s http://localhost:4500/README.md

Running 10s test @ http://localhost:4500/README.md
  8 threads and 20 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   149.85us   67.68us   6.95ms   96.22%
    Req/Sec    13.17k   555.45    14.67k    97.40%
  1057363 requests in 10.10s, 110.92MB read
Requests/sec: 104697.21
Transfer/sec:     10.98MB
```

this PR:

```
Running 10s test @ http://localhost:4500/README.md
  8 threads and 20 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   146.98us   62.17us   6.45ms   95.38%
    Req/Sec    13.41k   539.74    13.77k    99.01%
  1078083 requests in 10.10s, 113.10MB read
Requests/sec: 106747.55
Transfer/sec:     11.20MB
```